Revert "Feature/detail activity"
