--- app/src/main/java/udacity/lsaippa/cooking/network/api/ApiHelper.java
@@ -1,11 +1,0 @@
-package udacity.lsaippa.cooking.network.api;
-
-import java.util.List;
-
-import io.reactivex.Single;
-import udacity.lsaippa.cooking.network.model.Recipe;
-
-
-public interface ApiHelper {
-    Single<List<Recipe>> getRecipeApiCall();
-}
