--- app/src/main/java/udacity/lsaippa/cooking/network/api/RecipesApi.java
@@ -1,19 +1,0 @@
-package udacity.lsaippa.cooking.network.api;
-
-import java.util.List;
-
-import io.reactivex.Single;
-import retrofit2.http.GET;
-import udacity.lsaippa.cooking.network.model.Recipe;
-
-/**
- * Created by lsaippa on 12/01/19.
- */
-
-public interface RecipesApi {
-
-    @GET("topher/2017/May/59121517_baking/baking.json")
-    Single<List<Recipe>> getRecipes();
-
-
-}
