--- app/src/main/java/udacity/lsaippa/cooking/network/model/Ingredient.java
@@ -1,68 +1,0 @@
-package udacity.lsaippa.cooking.network.model;
-
-import android.os.Parcel;
-import android.os.Parcelable;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Ingredient implements Parcelable {
-    @SerializedName("quantity")
-    @Expose
-    private float quantity;
-    @SerializedName("measure")
-    @Expose
-    private String measure;
-    @SerializedName("ingredient")
-    @Expose
-    private String ingredient;
-
-    public float getQuantity() {
-        return quantity;
-    }
-
-
-    public String getMeasure() {
-        return measure;
-    }
-
-
-    public String getIngredient() {
-        return ingredient;
-    }
-
-
-
-    @Override
-    public int describeContents() {
-        return 0;
-    }
-
-    @Override
-    public void writeToParcel(Parcel dest, int flags) {
-        dest.writeFloat(this.quantity);
-        dest.writeString(this.measure);
-        dest.writeString(this.ingredient);
-    }
-
-    public Ingredient() {
-    }
-
-    protected Ingredient(Parcel in) {
-        this.quantity = in.readFloat();
-        this.measure = in.readString();
-        this.ingredient = in.readString();
-    }
-
-    public static final Parcelable.Creator<Ingredient> CREATOR = new Parcelable.Creator<Ingredient>() {
-        @Override
-        public Ingredient createFromParcel(Parcel source) {
-            return new Ingredient(source);
-        }
-
-        @Override
-        public Ingredient[] newArray(int size) {
-            return new Ingredient[size];
-        }
-    };
-}
