+++ app/src/main/java/udacity/lsaippa/cooking/network/model/Recipe.java
@@ -1,104 +1,4 @@
 package udacity.lsaippa.cooking.network.model;
 
-import android.os.Parcel;
-import android.os.Parcelable;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Recipe implements Parcelable{
-    @SerializedName("id")
-    @Expose
-    private int id;
-    @SerializedName("name")
-    @Expose
-    private String name;
-    @SerializedName("ingredients")
-    @Expose
-    private List<Ingredient> ingredients = null;
-    @SerializedName("steps")
-    @Expose
-    private ArrayList<Step> steps = null;
-    @SerializedName("servings")
-    @Expose
-    private int servings;
-    @SerializedName("image")
-    @Expose
-    private String image;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public List<Ingredient> getIngredients() {
-        return ingredients;
-    }
-
-    public ArrayList<Step> getSteps() {
-        return steps;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-
-
-    @Override
-    public int describeContents() {
-        return 0;
-    }
-
-    @Override
-    public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(this.id);
-        dest.writeString(this.name);
-        dest.writeList(this.ingredients);
-        dest.writeTypedList(this.steps);
-        dest.writeInt(this.servings);
-        dest.writeString(this.image);
-    }
-
-    public Recipe() {
-    }
-
-    protected Recipe(Parcel in) {
-        this.id = in.readInt();
-        this.name = in.readString();
-        this.ingredients = new ArrayList<>();
-        in.readList(this.ingredients, Ingredient.class.getClassLoader());
-        this.steps = in.createTypedArrayList(Step.CREATOR);
-        this.servings = in.readInt();
-        this.image = in.readString();
-    }
-
-    public static final Creator<Recipe> CREATOR = new Creator<Recipe>() {
-        @Override
-        public Recipe createFromParcel(Parcel source) {
-            return new Recipe(source);
-        }
-
-        @Override
-        public Recipe[] newArray(int size) {
-            return new Recipe[size];
-        }
-    };
+public class Recipe {
 }
