--- app/src/main/java/udacity/lsaippa/cooking/utils/AppConstants.java
@@ -1,14 +1,0 @@
-package udacity.lsaippa.cooking.utils;
-
-import udacity.lsaippa.cooking.network.model.Recipe;
-
-/**
- * Created by lsaippa on 12/01/19.
- */
-
-public class AppConstants {
-
-    public static final String BASE_URL = "https://d17h27t6h515a5.cloudfront.net/";
-
-    public static final String RECIPE_TAG = Recipe.class.getSimpleName();
-}
