--- app/src/main/res/layout/detail_recipe_steps_view_holder.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-android:layout_width="match_parent"
-android:layout_height="wrap_content"
-android:orientation="horizontal">
-
-<TextView
-    android:id="@+id/tv_step_name"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:textStyle="bold"
-    />
-
-
-<ImageView
-    android:id="@+id/imv_step_thumbnail"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:layout_gravity="end"
-    android:src="@android:drawable/ic_media_play"
-    />
-</FrameLayout>
-
-
