--- app/src/main/res/layout/frg_recipe_details.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
-    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/rv_detail"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"/>
-
-</FrameLayout>
